Issue #15: Add method to create location details
This can be called in prepareForSegue so that we can pass the location
details back to the location view.

The call to this method is currently in the button 
alertView clickedButtonAtIndex delegate method simply to check it works
as expected. It is envisioned that this delegate method with simply call
the prepareForSegue method and nothing else.