Issue #63: Added script to populate ElasticSearch index.
To modify the location of the estates.json file or ElasticSearch host and port, modify the script directly. Currently it is pointing to an install on my local machine - this will have to be changed in future.

You will also need to install the elastic search module:

‘pip install elasticsearch’ 

To run:

python populateElasticSearch.py