Issue #96: Added image orientation fix.
Note that this is a little clunky - we have to read the exif data from images when loading with ALAssetsLibrary.  Dependent on the value for ‘Orientation’, we have to create a new image that has been rotated manually. We do this twice - once in ViewPhotoViewController so that when the user views the image it is rotated correctly and once in SummaryViewController to ensure we post the image with the correct exif data so that it can be viewed in the right orientation in JitBit.  We have to do it this way as image.imageOrientation always returns UIImageOrientationUp no matter what way the photo was taken - possibly it’s because we are using UIImagePickerController.  The only way to get the true orientation is to read the exif data directly.